moved some reusable code from hysr_control.py to config.py
